+++ stores/entities/store_photo.go
@@ -1,7 +1,7 @@
 package entities
 
 type StorePhoto struct {
-	ID        uint    `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID        int64   `gorm:"primaryKey;autoIncrement" json:"id"`
 	StoreID   *uint   `gorm:"column:storeId;index:idx_store_photo_type" json:"storeId"`
 	Type      string  `gorm:"type:enum('cover','sign','inside','roadview')" json:"type"`
 	ImageName *string `gorm:"column:imageName;type:varchar(255)" json:"imageName"`
